CDSR-650: rename DeclarationDetailsAnswer model
