+++ app/uk/gov/hmrc/cdsreimbursementclaimfrontend/models/DraftClaim.scala
@@ -38,29 +38,29 @@ sealed trait DraftClaim extends Product with Serializable {
 object DraftClaim {
 
   final case class DraftC285Claim(
-    id: UUID,
-    selectNumberOfClaimsAnswer: Option[SelectNumberOfClaimsAnswer],
-    movementReferenceNumber: Option[MovementReferenceNumber],
-    duplicateMovementReferenceNumberAnswer: Option[MovementReferenceNumber],
-    declarationDetailsAnswer: Option[DeclarationDetailsAnswer],
-    duplicateDeclarationDetailsAnswer: Option[DuplicateDeclarationDetailsAnswer],
-    declarantTypeAnswer: Option[DeclarantTypeAnswer],
-    detailsRegisteredWithCdsAnswer: Option[DetailsRegisteredWithCdsAnswer],
-    contactDetailsAnswer: Option[ContactDetailsAnswer],
-    bankAccountDetailsAnswer: Option[BankAccountDetailsAnswer],
-    basisOfClaimAnswer: Option[BasisOfClaimAnswer],
-    supportingEvidenceAnswer: Option[SupportingEvidenceAnswer],
-    dutiesSelectedAnswer: Option[DutiesSelectedAnswer],
-    commoditiesDetailsAnswer: Option[CommodityDetails],
-    claimNorthernIrelandAnswer: Option[ClaimNorthernIrelandAnswer],
-    reasonForBasisAndClaimAnswer: Option[ReasonAndBasisOfClaimAnswer],
-    displayDeclaration: Option[DisplayDeclaration],
-    duplicateDisplayDeclaration: Option[DisplayDeclaration],
-    importerEoriNumberAnswer: Option[ImporterEoriNumberAnswer],
-    declarantEoriNumberAnswer: Option[DeclarantEoriNumberAnswer],
-    claimsAnswer: Option[ClaimsAnswer],
-    checkClaimAnswer: Option[CheckClaimAnswer],
-    checkDeclarationDetailsAnswer: Option[CheckDeclarationDetailsAnswer]
+                                   id: UUID,
+                                   selectNumberOfClaimsAnswer: Option[SelectNumberOfClaimsAnswer],
+                                   movementReferenceNumber: Option[MovementReferenceNumber],
+                                   duplicateMovementReferenceNumberAnswer: Option[MovementReferenceNumber],
+                                   declarationDetailsAnswer: Option[DeclarationDetails],
+                                   duplicateDeclarationDetailsAnswer: Option[DuplicateDeclarationDetailsAnswer],
+                                   declarantTypeAnswer: Option[DeclarantTypeAnswer],
+                                   detailsRegisteredWithCdsAnswer: Option[DetailsRegisteredWithCdsAnswer],
+                                   contactDetailsAnswer: Option[ContactDetailsAnswer],
+                                   bankAccountDetailsAnswer: Option[BankAccountDetailsAnswer],
+                                   basisOfClaimAnswer: Option[BasisOfClaimAnswer],
+                                   supportingEvidenceAnswer: Option[SupportingEvidenceAnswer],
+                                   dutiesSelectedAnswer: Option[DutiesSelectedAnswer],
+                                   commoditiesDetailsAnswer: Option[CommodityDetails],
+                                   claimNorthernIrelandAnswer: Option[ClaimNorthernIrelandAnswer],
+                                   reasonForBasisAndClaimAnswer: Option[ReasonAndBasisOfClaimAnswer],
+                                   displayDeclaration: Option[DisplayDeclaration],
+                                   duplicateDisplayDeclaration: Option[DisplayDeclaration],
+                                   importerEoriNumberAnswer: Option[ImporterEoriNumberAnswer],
+                                   declarantEoriNumberAnswer: Option[DeclarantEoriNumberAnswer],
+                                   claimsAnswer: Option[ClaimsAnswer],
+                                   checkClaimAnswer: Option[CheckClaimAnswer],
+                                   checkDeclarationDetailsAnswer: Option[CheckDeclarationDetailsAnswer]
   ) extends DraftClaim
 
   object DraftC285Claim {
